refactor: unify account/nullifier SMT backends into shared SmtBackend traits
Replaces the near-identical AccountTreeBackend/NullifierTreeBackend trait
pairs with a single Word-valued SmtBackendReader + SmtBackend pair, impl'd
once for Smt and once for LargeSmt. The NullifierBlock value wrapping moves
from the backend into NullifierTree's own accessors, leaving the backend
layer value-agnostic. Removes both per-tree backend.rs modules.
